+++ src/components/dropdown-filter/dropdown-filter.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import Dropdown from './../dropdown';
 import I18n from 'i18n-js';
+import escapeStringRegexp from 'escape-string-regexp';
 
 /**
  * A dropdown filter widget.
@@ -130,10 +131,10 @@ class DropdownFilter extends Dropdown {
       highlighted: null
     };
 
-    if (this.props.suggest && ev.target.value.length > 0) {
-      state.open = true;
-    } else if (this.props.suggest) {
+    if (this.props.suggest && ev.target.value.length <= 0) {
       state.open = false;
+    } else {
+      state.open = true;
     }
 
     this.setState(state);
@@ -156,7 +157,7 @@ class DropdownFilter extends Dropdown {
       let filter = this.props.create ? this.state.filter : null,
           highlighted = this.highlighted(this.options);
 
-      if (highlighted != this.props.value) {
+      if (highlighted && highlighted !== String(this.props.value)) {
         let item = this.props.options.find((item) => {
           return String(item.get('id')) === String(highlighted);
         });
@@ -174,11 +175,13 @@ class DropdownFilter extends Dropdown {
    * @method handleFocus
    */
   handleFocus = () => {
-    if (!this.props.suggest) {
+    if (!this.props.suggest && !this.blockFocus) {
       this.setState({ open: true });
+    } else {
+      this.blockFocus = false;
     }
 
-    this.refs.input.setSelectionRange(0, this.refs.input.value.length);
+    this._input.setSelectionRange(0, this._input.value.length);
   }
 
   /**
@@ -200,7 +203,7 @@ class DropdownFilter extends Dropdown {
   prepareList = (options) => {
     if ((this.props.suggest || !this.openingList) && typeof this.state.filter === 'string') {
       let filter = this.state.filter;
-      let regex = new RegExp(filter, 'i');
+      let regex = new RegExp(escapeStringRegexp(filter), 'i');
 
       // if user has entered a search filter
       options = options.filter((option) => {