+++ src/components/dropdown/__spec__.js
@@ -2,6 +2,7 @@ import React from 'react';
 import TestUtils from 'react/lib/ReactTestUtils';
 import Dropdown from './dropdown';
 import Immutable from 'immutable';
+import Events from './../../utils/helpers/events';
 
 describe('Dropdown', () => {
   let instance;
@@ -42,6 +43,25 @@ describe('Dropdown', () => {
           value: "1"
         });
         expect(instance.visibleValue).toBe("foobar");
+      });
+    });
+  });
+
+  describe('componentDidMount', () => {
+    describe('if not autoFocus', () => {
+      it('does not set focus on the input', () => {
+        spyOn(instance._input, 'focus');
+        instance.componentDidMount();
+        expect(instance._input.focus).not.toHaveBeenCalled();
+      });
+    });
+
+    describe('if autoFocus', () => {
+      it('does sets focus on the input', () => {
+        instance = TestUtils.renderIntoDocument(<Dropdown options={ Immutable.fromJS([{}]) } autoFocus />);
+        spyOn(instance._input, 'focus');
+        instance.componentDidMount();
+        expect(instance._input.focus).toHaveBeenCalled();
       });
     });
   });
@@ -149,7 +169,7 @@ describe('Dropdown', () => {
 
   describe('handleMouseDownOnList', () => {
     beforeEach(() => {
-      spyOn(instance.refs.input, 'focus');
+      spyOn(instance._input, 'focus');
       jasmine.clock().install();
     });
 
@@ -163,7 +183,7 @@ describe('Dropdown', () => {
           target: instance.refs.list
         });
         jasmine.clock().tick();
-        expect(instance.refs.input.focus).toHaveBeenCalled();
+        expect(instance._input.focus).toHaveBeenCalled();
       });
     });
 
@@ -173,7 +193,7 @@ describe('Dropdown', () => {
           target: 'foo'
         });
         jasmine.clock().tick();
-        expect(instance.refs.input.focus).not.toHaveBeenCalled();
+        expect(instance._input.focus).not.toHaveBeenCalled();
       });
     });
   });
@@ -186,7 +206,7 @@ describe('Dropdown', () => {
     describe('if blur is blocked', () => {
       it('does not call setState', () => {
         instance.blockBlur = true;
-        TestUtils.Simulate.blur(instance.refs.input);
+        TestUtils.Simulate.blur(instance._input);
         expect(instance.setState).not.toHaveBeenCalled();
       });
     });
@@ -202,14 +222,14 @@ describe('Dropdown', () => {
 
       it('calls setState', () => {
         instance.blockBlur = false;
-        TestUtils.Simulate.blur(instance.refs.input);
+        TestUtils.Simulate.blur(instance._input);
         expect(instance.setState).toHaveBeenCalledWith({ open: false });
       });
 
       describe('if highlighted matches value', () => {
         it('does not emit change', () => {
           spyOn(instance, 'highlighted').and.returnValue(instance.props.value);
-          TestUtils.Simulate.blur(instance.refs.input);
+          TestUtils.Simulate.blur(instance._input);
           expect(instance.emitOnChangeCallback).not.toHaveBeenCalled();
         });
       });
@@ -217,7 +237,7 @@ describe('Dropdown', () => {
       describe('if highlighted does not match value', () => {
         it('emits change', () => {
           spyOn(instance, 'highlighted').and.returnValue('2');
-          TestUtils.Simulate.blur(instance.refs.input);
+          TestUtils.Simulate.blur(instance._input);
           expect(instance.emitOnChangeCallback).toHaveBeenCalledWith('2', 'bar');
         });
       });
@@ -225,11 +245,22 @@ describe('Dropdown', () => {
   });
 
   describe('handleFocus', () => {
-    it('calls setState', () => {
-      spyOn(instance, 'setState');
-      TestUtils.Simulate.focus(instance.refs.input);
-      expect(instance.setState).toHaveBeenCalledWith({
-        open: true
+    describe('if focus is not blocked', () => {
+      it('calls setState', () => {
+        spyOn(instance, 'setState');
+        TestUtils.Simulate.focus(instance._input);
+        expect(instance.setState).toHaveBeenCalledWith({
+          open: true
+        });
+      });
+    });
+
+    describe('if focus is blocked', () => {
+      it('does not call setState', () => {
+        instance.blockFocus = true;
+        spyOn(instance, 'setState');
+        TestUtils.Simulate.focus(instance._input);
+        expect(instance.setState).not.toHaveBeenCalled();
       });
     });
 
@@ -239,7 +270,7 @@ describe('Dropdown', () => {
           <Dropdown name="foo" options={ Immutable.fromJS([{ id: 1, name: 'foo' }]) } readOnly={ true } />
         );
         spyOn(instance, 'setState');
-        TestUtils.Simulate.focus(instance.refs.input);
+        TestUtils.Simulate.focus(instance._input);
         expect(instance.setState).not.toHaveBeenCalled();
       });
     });
@@ -250,7 +281,7 @@ describe('Dropdown', () => {
           <Dropdown name="foo" options={ Immutable.fromJS([{ id: 1, name: 'foo' }]) } disabled={ true } />
         );
         spyOn(instance, 'setState');
-        TestUtils.Simulate.focus(instance.refs.input);
+        TestUtils.Simulate.focus(instance._input);
         expect(instance.setState).not.toHaveBeenCalled();
       });
     });
@@ -296,15 +327,54 @@ describe('Dropdown', () => {
   });
 
   describe('handleKeyDown', () => {
+    let ev;
+
     beforeEach(() => {
+      ev = { which: "" };
+
       instance = TestUtils.renderIntoDocument(
         <Dropdown name="foo" options={ Immutable.fromJS([{ id: 1, name: 'foo' }, { id: 2, name: 'bar' }]) } value="" />
       );
     });
 
     describe('if there is no list', () => {
-      it('returns nothing', () => {
-        expect(instance.handleKeyDown()).toBe(undefined);
+      beforeEach(() => {
+        spyOn(instance, 'setState');
+      });
+
+      describe('if up key', () => {
+        it('opens the list', () => {
+          spyOn(Events, 'isUpKey').and.returnValue(true);
+          let result = instance.handleKeyDown(ev);
+          expect(instance.setState).toHaveBeenCalledWith({ open: true });
+          expect(result).toBe(undefined);
+        });
+      });
+
+      describe('if down key', () => {
+        it('opens the list', () => {
+          spyOn(Events, 'isDownKey').and.returnValue(true);
+          let result = instance.handleKeyDown(ev);
+          expect(instance.setState).toHaveBeenCalledWith({ open: true });
+          expect(result).toBe(undefined);
+        });
+      });
+
+      describe('if space key', () => {
+        it('opens the list', () => {
+          spyOn(Events, 'isSpaceKey').and.returnValue(true);
+          let result = instance.handleKeyDown(ev);
+          expect(instance.setState).toHaveBeenCalledWith({ open: true });
+          expect(result).toBe(undefined);
+        });
+      });
+
+      describe('if other key', () => {
+        it('does not open list, but still returns undefined', () => {
+          let result = instance.handleKeyDown(ev);
+          expect(instance.setState).not.toHaveBeenCalled();
+          expect(result).toBe(undefined);
+        });
       });
     });
 
@@ -325,7 +395,7 @@ describe('Dropdown', () => {
         describe('if something is highlighted', () => {
           beforeEach(() => {
             instance.setState({ highlighted: 1 });
-            TestUtils.Simulate.keyDown(instance.refs.input, opts);
+            TestUtils.Simulate.keyDown(instance._input, opts);
           });
 
           it('prevents default', () => {
@@ -340,7 +410,7 @@ describe('Dropdown', () => {
         describe('if something is not highlighted', () => {
           it('does not prevent default', () => {
             instance.setState({ highlighted: 'abc' }); // value which does not exist
-            TestUtils.Simulate.keyDown(instance.refs.input, opts);
+            TestUtils.Simulate.keyDown(instance._input, opts);
             expect(spy).not.toHaveBeenCalled();
           });
         });
@@ -355,7 +425,7 @@ describe('Dropdown', () => {
         });
 
         it('prevents default', () => {
-          TestUtils.Simulate.keyDown(instance.refs.input, opts);
+          TestUtils.Simulate.keyDown(instance._input, opts);
           expect(spy).toHaveBeenCalled();
         });
 
@@ -363,8 +433,8 @@ describe('Dropdown', () => {
           it('calls setState with the correct values', () => {
             instance.setState({ highlighted: 1 });
             spyOn(instance, 'setState');
-            TestUtils.Simulate.keyDown(instance.refs.input, opts);
-            expect(instance.setState).toHaveBeenCalledWith({ highlighted: 2 });
+            TestUtils.Simulate.keyDown(instance._input, opts);
+            expect(instance.setState).toHaveBeenCalledWith({ highlighted: '2' });
           });
         });
 
@@ -372,8 +442,8 @@ describe('Dropdown', () => {
           it('calls setState with the correct values', () => {
             instance.setState({ highlighted: 2 });
             spyOn(instance, 'setState');
-            TestUtils.Simulate.keyDown(instance.refs.input, opts);
-            expect(instance.setState).toHaveBeenCalledWith({ highlighted: 1 });
+            TestUtils.Simulate.keyDown(instance._input, opts);
+            expect(instance.setState).toHaveBeenCalledWith({ highlighted: '1' });
           });
         });
       });
@@ -387,7 +457,7 @@ describe('Dropdown', () => {
         });
 
         it('prevents default', () => {
-          TestUtils.Simulate.keyDown(instance.refs.input, opts);
+          TestUtils.Simulate.keyDown(instance._input, opts);
           expect(spy).toHaveBeenCalled();
         });
 
@@ -395,8 +465,8 @@ describe('Dropdown', () => {
           it('calls setState with the correct values', () => {
             instance.setState({ highlighted: 2 });
             spyOn(instance, 'setState');
-            TestUtils.Simulate.keyDown(instance.refs.input, opts);
-            expect(instance.setState).toHaveBeenCalledWith({ highlighted: 1 });
+            TestUtils.Simulate.keyDown(instance._input, opts);
+            expect(instance.setState).toHaveBeenCalledWith({ highlighted: '1' });
           });
         });
 
@@ -404,8 +474,23 @@ describe('Dropdown', () => {
           it('calls setState with the correct values', () => {
             instance.setState({ highlighted: 1 });
             spyOn(instance, 'setState');
-            TestUtils.Simulate.keyDown(instance.refs.input, opts);
-            expect(instance.setState).toHaveBeenCalledWith({ highlighted: 2 });
+            TestUtils.Simulate.keyDown(instance._input, opts);
+            expect(instance.setState).toHaveBeenCalledWith({ highlighted: '2' });
+          });
+        });
+
+        describe('when strings are used for IDs', () => {
+          beforeEach(() => {
+            instance = TestUtils.renderIntoDocument(
+              <Dropdown name="foo" options={ Immutable.fromJS([{ id: 'foo', name: 'foo' }, { id: 'bar', name: 'bar' }]) } value="" />
+            );
+            instance.setState({ open: true });
+          });
+
+          it('still works', () => {
+            spyOn(instance, 'setState');
+            TestUtils.Simulate.keyDown(instance._input, opts);
+            expect(instance.setState).toHaveBeenCalledWith({ highlighted: 'foo' });
           });
         });
       });
@@ -458,7 +543,6 @@ describe('Dropdown', () => {
       instance.visibleValue = 'foo';
       expect(instance.inputProps.className).toEqual(instance.inputClasses);
       expect(instance.inputProps.name).toBe(null);
-      expect(instance.inputProps.ref).toEqual("input");
       expect(instance.inputProps.readOnly).toBeTruthy();
       expect(instance.inputProps.value).toEqual('foo');
     });