+++ src/components/dropdown/dropdown.js
@@ -3,6 +3,7 @@ import Input from './../../utils/decorators/input';
 import InputLabel from './../../utils/decorators/input-label';
 import InputValidation from './../../utils/decorators/input-validation';
 import InputIcon from './../../utils/decorators/input-icon';
+import Events from './../../utils/helpers/events';
 
 /**
  * A dropdown widget.
@@ -103,6 +104,18 @@ class Dropdown extends React.Component {
   }
 
   /**
+   * Manually focus if autoFocus is applied - allows us to prevent the list from opening.
+   *
+   * @method componentDidMount
+   */
+  componentDidMount() {
+    if (this.props.autoFocus) {
+      this.blockFocus = true;
+      this._input.focus();
+    }
+  }
+
+  /**
    * Clears the visible value if a new value has been selected.
    *
    * @method componentWillReceiveProps
@@ -194,7 +207,7 @@ class Dropdown extends React.Component {
     // NOTE: this is an IE11 fix
     if (ev.target === this.refs.list) {
       setTimeout(() => {
-        this.refs.input.focus();
+        this._input.focus();
       }, 0);
     }
   }
@@ -226,9 +239,11 @@ class Dropdown extends React.Component {
    * @method handleFocus
    */
   handleFocus = () => {
-    this.setState({
-      open: true
-    });
+    if (this.blockFocus) {
+      this.blockFocus = false;
+    } else {
+      this.setState({ open: true });
+    }
   }
 
   /**
@@ -264,7 +279,14 @@ class Dropdown extends React.Component {
    * @param {Object} ev event
    */
   handleKeyDown = (ev) => {
-    if (!this.refs.list) { return; }
+    if (!this.refs.list) {
+      // if up/down/space then open list
+      if (Events.isUpKey(ev) || Events.isDownKey(ev) || Events.isSpaceKey(ev)) {
+        this.setState({ open: true });
+      }
+
+      return;
+    }
 
     let list = this.refs.list,
         element = list.getElementsByClassName('ui-dropdown__list__item--highlighted')[0],
@@ -279,20 +301,20 @@ class Dropdown extends React.Component {
         break;
       case 38: // up arrow
         ev.preventDefault();
-        nextVal = list.lastChild.value;
+        nextVal = list.lastChild.getAttribute('value');
 
         if (element && element.previousElementSibling) {
-          nextVal = element.previousElementSibling.value;
+          nextVal = element.previousElementSibling.getAttribute('value');
         }
 
         this.setState({ highlighted: nextVal });
         break;
       case 40: // down arrow
         ev.preventDefault();
-        nextVal = list.firstChild.value;
+        nextVal = list.firstChild.getAttribute('value');
 
         if (element && element.nextElementSibling) {
-          nextVal = element.nextElementSibling.value;
+          nextVal = element.nextElementSibling.getAttribute('value');
         }
 
         this.setState({ highlighted: nextVal });
@@ -335,13 +357,12 @@ class Dropdown extends React.Component {
    * @method inputProps
    */
   get inputProps() {
-    let { ...props } = this.props;
+    let { autoFocus, ...props } = this.props;
     props.className = this.inputClasses;
     props.value = this.visibleValue || this.nameByID();
     props.name = null;
     props.onBlur = this.handleBlur;
     props.onKeyDown = this.handleKeyDown;
-    props.ref = "input";
     props.readOnly = true;
 
     if (!this.props.readOnly && !this.props.disabled) {