+++ src/utils/decorators/input-validation/__spec__.js
@@ -353,6 +353,19 @@ describe('InputValidation', () => {
           expect(validationThree.validate).toHaveBeenCalledWith(instance.props.value, instance.props);
         });
 
+        describe('when the second validation fails', () => {
+          it('stops validating', () => {
+            instance = TestUtils.renderIntoDocument(React.createElement(Component, {
+              validations: [validationOne, validationThree, validationTwo],
+              value: 'foo'
+            }));
+            instance.validate();
+            expect(validationOne.validate).toHaveBeenCalledWith(instance.props.value, instance.props);
+            expect(validationThree.validate).toHaveBeenCalledWith(instance.props.value, instance.props);
+            expect(validationTwo.validate).not.toHaveBeenCalled();
+          });
+        });
+
         describe('when called with a custom value', () => {
           it('calls validate for each validation', () => {
             instance.validate('foo');
@@ -415,9 +428,18 @@ describe('InputValidation', () => {
   });
 
   describe('_handleBlur', () => {
+    beforeEach(() => {
+      jasmine.clock().install();
+    });
+
+    afterEach(() => {
+      jasmine.clock().uninstall();
+    });
+
     it('calls validate on blur of the input', () => {
       spyOn(instance, 'validate');
       instance._handleBlur();
+      jasmine.clock().tick(0);
       expect(instance.validate).toHaveBeenCalled();
     });
 
@@ -426,6 +448,7 @@ describe('InputValidation', () => {
         instance.setState({ messageLocked: true });
         spyOn(instance, 'setState');
         instance._handleBlur();
+        jasmine.clock().tick(0);
         expect(instance.setState).toHaveBeenCalledWith({ messageLocked: false });
       });
     });